Update PluginInterface to standard, adjust scintillaGateway in MarkdownPanelController.cs
When switching buffers or new instantiation of visible window, the Scintilla
component could be in the "other" view.  So we cannot have `scintillaGateway`
be `readonly`.  Previously, this was adjusted by adding a non-`readonly`
variable and SetScintillaHandle() in PluginInfrastructure/ScintillaGateway.cs.
This appears to be non-standard as the updated PluginInfrastructure from
@mohzy83 does not include this.  Instead, we adjust the now non-`readonly`
variable in MarkdownPanelController.cs and use the updated routine:

```csharp
        protected void UpdateEditorInformation()
        {
            scintillaGateway = new ScintillaGateway(PluginBase.GetCurrentScintilla());
        }
```