Ignore "invalid" code block languages
It seems Python isn't inluded, which is tragic.  But if we use:

```python
```

it will render just a plain text, not even a code block.  The colorer must
ignore what it thinks are "invalid" languages and just render the default
code block.  This commit does the language type check before the coloring
so it can bail if need be.  It still does the check again in the coloring,
we could get more efficient, but alas, this is not the optimal solution
anyway.