Handle unknown code fence labels
Bring this inline with a PR by @rdipardo in @mohzy83 master to fix issue.

From previous implementation:
It seems Python isn't inluded, which is tragic.  But if we use:

    ```python
    ```
it will render just a plain text, not even a code block.  The colorer must
ignore what it thinks are "invalid" languages and just render the default
code block.